Reorganize deleteCollection
