+++ tests/src/deleteCollections.test.ts
@@ -1,30 +1,36 @@
-import { createCollection } from "./util/createCollection";
+import { getApiConnection } from "./substrate/substrate-api";
 import {
+  createCollection,
   deleteCollection,
-  negativeDeleteCollection,
-} from "./util/deleteCollection";
+  negativeDeleteCollection
+} from "./util/tx";
 
 describe("Integration test: delete collection", () => {
+  let api: any;
+  before(async () => { api = await getApiConnection(); });
+
   const Alice = "//Alice";
   const Bob = "//Bob";
 
   it("Delete NFT collection", async () => {
-    await createCollection(Alice, "test-metadata", null, "test-symbol").then(
+    await createCollection(api, Alice, "test-metadata", null, "test-symbol").then(
       async (collectionId) => {
-        await deleteCollection(Alice, collectionId.toString());
+        await deleteCollection(api, Alice, collectionId.toString());
       }
     );
   });
 
-  it("[Negative] Delete non-existing NFT collection", async () => {
-    await negativeDeleteCollection(Alice, "99999");
+  it("[negative] delete non-existing NFT collection", async () => {
+    await negativeDeleteCollection(api, Alice, "99999");
   });
 
-  it("[Negative] Delete not an owner NFT collection", async () => {
-    await createCollection(Alice, "test-metadata", null, "test-symbol").then(
+  it("[negative] delete not an owner NFT collection", async () => {
+    await createCollection(api, Alice, "test-metadata", null, "test-symbol").then(
       async (collectionId) => {
-        await negativeDeleteCollection(Bob, collectionId.toString());
+        await negativeDeleteCollection(api, Bob, collectionId.toString());
       }
     );
   });
+
+  after(() => { api.disconnect(); });
 });
